+++ sticky/__init__.py
@@ -1,0 +1,3 @@
+"""Tacklet package metadata."""
+
+__version__ = '0.2.0'